BootMiss-class {semTools}	R Documentation

Class For the Results of Bollen-Stine Bootstrap with Incomplete Data

Description

This class contains the results of Bollen-Stine bootstrap with missing data.

Usage

## S4 method for signature 'BootMiss'
show(object)

## S4 method for signature 'BootMiss'
summary(object)

## S4 method for signature 'BootMiss'
hist(x, ..., alpha = 0.05, nd = 2,
  printLegend = TRUE, legendArgs = list(x = "topleft"))

Arguments

object, x
object of class BootMiss

...
Additional arguments to pass to graphics::hist()

alpha
alpha level used to draw confidence limits

nd
number of digits to display

printLegend
logical. If TRUE (default), a legend will be printed with the histogram

legendArgs
list of arguments passed to the graphics::legend() function. The default argument is a list placing the legend at the top-left of the figure.

Value

The hist method returns a list of length == 2, containing the arguments for the call to hist and the arguments to the call for legend, respectively.

Slots

time

A list containing 2 difftime objects (transform and fit), indicating the time elapsed for data transformation and for fitting the model to bootstrap data sets, respectively.

transData

Transformed data

bootDist

The vector of chi^2 values from bootstrap data sets fitted by the target model

origChi

The chi^2 value from the original data set

df

The degree of freedom of the model

bootP

The p value comparing the original chi^2 with the bootstrap distribution

Objects from the Class

Objects can be created via the bsBootMiss() function.

Author(s)

Terrence D. Jorgensen (University of Amsterdam; TJorgensen314@gmail.com)

See Also

bsBootMiss()

Examples

# See the example from the bsBootMiss function

[Package semTools version 0.5-8 Index]