Net-class {semTools}	R Documentation

Class For the Result of Nesting and Equivalence Testing

Description

This class contains the results of nesting and equivalence testing among multiple models

Usage

## S4 method for signature 'Net'
show(object)

## S4 method for signature 'Net'
summary(object)

Arguments

object
An object of class Net.

Value

show
signature(object = "Net"): prints the logical matrix of test results. NA indicates a model did not converge.

summary
signature(object = "Net"): prints a narrative description of results. The original object is invisibly returned.

Slots

test

Logical matrix indicating nesting/equivalence among models

df

The degrees of freedom of tested models

Objects from the Class

Objects can be created via the net() function.

Author(s)

Terrence D. Jorgensen (University of Amsterdam; TJorgensen314@gmail.com)

See Also

net()

Examples

# See the example in the net function.

[Package semTools version 0.5-8 Index]