clipboard {semTools}	R Documentation

Copy or save the result of lavaan or FitDiff objects into a clipboard or a file

Description

Copy or save the result of lavaan or FitDiff object into a clipboard or a file. From the clipboard, users may paste the result into the Microsoft Excel or spreadsheet application to create a table of the output.

Usage

clipboard(object, what = "summary", ...)

saveFile(object, file, what = "summary", tableFormat = FALSE,
  fit.measures = "default", writeArgs = list(), ...)

Arguments

object
An object of class lavaan::lavaan or FitDiff.

what
The attributes of the lavaan object to be copied in the clipboard. "summary" is to copy the screen provided from the summary function. "epceqfit" is to copy the result from the epcEquivFit() function. Other attributes listed in the inspect method in the lavaan::lavaan could also be used, such as "coef", "se", "fit", "samp", and so on. Ignored for FitDiff-class objects.

...
Additional arguments when passing a lavaan object to the summary or epcEquivFit() function.

file
A file name used for saving the result.

tableFormat
If TRUE, save the result in the table format using tabs for separation. Otherwise, save the result as the output screen printed in the R console.

fit.measures
character vector specifying names of fit measures returned by lavaan::fitMeasures() to be copied/saved. Only relevant if object is class FitDiff.

writeArgs
list of additional arguments to be passed to utils::write.table()

Value

The resulting output will be saved into a clipboard or a file. If using the clipboard function, users may paste it in the other applications.

Author(s)

Sunthud Pornprasertmanit (psunthud@gmail.com)

Terrence D. Jorgensen (University of Amsterdam; TJorgensen314@gmail.com)

Examples

library(lavaan)
HW.model <- ' visual  =~ x1 + c1*x2 + x3
              textual =~ x4 + c1*x5 + x6
              speed   =~ x7 +    x8 + x9 '

fit <- cfa(HW.model, data = HolzingerSwineford1939, group = "school")

if(interactive()){
# Copy the summary of the lavaan object
clipboard(fit)

# pass additional arguments to summary() method for class?lavaan
clipboard(fit, rsquare = TRUE, standardized = TRUE, fit.measures = TRUE)

# Copy the EPC equivalence testing results from the epcEquivFit() function
clipboard(fit, "epceqfit")

# Copy the parameter estimates
clipboard(fit, "coef")

# Copy the standard errors
clipboard(fit, "se")

# Copy the sample statistics
clipboard(fit, "samp")

# Copy the fit measures
clipboard(fit, "fit")

# Save the summary of the lavaan object
saveFile(fit, "out.txt")

# Save the EPC equivalence testing results from the epcEquivFit() function
saveFile(fit, "out.txt", "epceqfit")

# Save the parameter estimates
saveFile(fit, "out.txt", "coef")

# Save the standard errors
saveFile(fit, "out.txt", "se")

# Save the sample statistics
saveFile(fit, "out.txt", "samp")

# Save the fit measures
saveFile(fit, "out.txt", "fit")
}

[Package semTools version 0.5-8 Index]