findRMSEApowernested {semTools}	R Documentation

Find power given a sample size in nested model comparison

Description

Find the sample size that the power in rejection the samples from the alternative pair of RMSEA is just over the specified power.

Usage

findRMSEApowernested(rmsea0A = NULL, rmsea0B = NULL, rmsea1A,
  rmsea1B = NULL, dfA, dfB, n, alpha = 0.05, group = 1)

Arguments

rmsea0A
The H_0 baseline RMSEA

rmsea0B
The H_0 alternative RMSEA (trivial misfit)

rmsea1A
The H_1 baseline RMSEA

rmsea1B
The H_1 alternative RMSEA (target misfit to be rejected)

dfA
degree of freedom of the more-restricted model

dfB
degree of freedom of the less-restricted model

n
Sample size

alpha
The alpha level

group
The number of group in calculating RMSEA

Author(s)

Bell Clinton

Pavel Panko (Texas Tech University; pavel.panko@ttu.edu)

Sunthud Pornprasertmanit (psunthud@gmail.com)

References

MacCallum, R. C., Browne, M. W., & Cai, L. (2006). Testing differences between nested covariance structure models: Power analysis and null hypotheses. Psychological Methods, 11(1), 19–35. doi: 10.1037/1082-989X.11.1.19

See Also

plotRMSEApowernested() to plot the statistical power for nested model comparison based on population RMSEA given the sample size

findRMSEAsamplesizenested() to find the minium sample size for a given statistical power in nested model comparison based on population RMSEA

Examples

findRMSEApowernested(rmsea0A = 0.06, rmsea0B = 0.05, rmsea1A = 0.08,
                     rmsea1B = 0.05, dfA = 22, dfB = 20, n = 200,
                     alpha = 0.05, group = 1)

[Package semTools version 0.5-8 Index]