kurtosis {semTools}	R Documentation

Finding excessive kurtosis

Description

Finding excessive kurtosis (g_{2}) of an object

Usage

kurtosis(object, population = FALSE)

Arguments

object
A vector used to find a excessive kurtosis

population
TRUE to compute the parameter formula. FALSE to compute the sample statistic formula.

Details

The excessive kurtosis computed by default is g_{2}, the fourth standardized moment of the empirical distribution of object. The population parameter excessive kurtosis γ_{2} formula is

γ_{2} = \frac{μ_{4}}{μ^{2}_{2}} - 3,

where μ_{i} denotes the i order central moment.

The excessive kurtosis formula for sample statistic g_{2} is

g_{2} = \frac{k_{4}}{k^{2}_{2}} - 3,

where k_{i} are the i order k-statistic.

The standard error of the excessive kurtosis is

Var(\hat{g}_{2}) = \frac{24}{N}

where N is the sample size.

Value

A value of an excessive kurtosis with a test statistic if the population is specified as FALSE

Author(s)

Sunthud Pornprasertmanit (psunthud@gmail.com)

References

Weisstein, Eric W. (n.d.). Kurtosis. Retrieved from MathWorld–A Wolfram Web Resource: https://mathworld.wolfram.com/Kurtosis.html

See Also

skew() Find the univariate skewness of a variable

mardiaSkew() Find the Mardia's multivariate skewness of a set of variables

mardiaKurtosis() Find the Mardia's multivariate kurtosis of a set of variables

Examples

kurtosis(1:5)

[Package semTools version 0.5-8 Index]