mardiaKurtosis {semTools}	R Documentation

Finding Mardia's multivariate kurtosis

Description

Finding Mardia's multivariate kurtosis of multiple variables

Usage

mardiaKurtosis(dat, use = "everything")

Arguments

dat
The target matrix or data frame with multiple variables

use
Missing data handling method from the stats::cov() function.

Details

The Mardia's multivariate kurtosis formula (Mardia, 1970) is

b_{2, d} = \frac{1}{n}∑^n_{i=1}≤ft[ ≤ft(\bold{X}_i - \bold{\bar{X}} \right)^{'} \bold{S}^{-1} ≤ft(\bold{X}_i - \bold{\bar{X}} \right) \right]^2,

where d is the number of variables, X is the target dataset with multiple variables, n is the sample size, \bold{S} is the sample covariance matrix of the target dataset, and \bold{\bar{X}} is the mean vectors of the target dataset binded in n rows. When the population multivariate kurtosis is normal, the b_{2,d} is asymptotically distributed as normal distribution with the mean of d(d + 2) and variance of 8d(d + 2)/n.

Value

A value of a Mardia's multivariate kurtosis with a test statistic

Author(s)

Sunthud Pornprasertmanit (psunthud@gmail.com)

References

Mardia, K. V. (1970). Measures of multivariate skewness and kurtosis with applications. Biometrika, 57(3), 519–530. doi: 10.2307/2334770

See Also

skew() Find the univariate skewness of a variable

kurtosis() Find the univariate excessive kurtosis of a variable

mardiaSkew() Find the Mardia's multivariate skewness of a set of variables

Examples

library(lavaan)
mardiaKurtosis(HolzingerSwineford1939[ , paste0("x", 1:9)])

[Package semTools version 0.5-8 Index]