plotRMSEAdist {semTools}	R Documentation

Plot the sampling distributions of RMSEA

Description

Plots the sampling distributions of RMSEA based on the noncentral chi-square distributions

Usage

plotRMSEAdist(rmsea, n, df, ptile = NULL, caption = NULL,
  rmseaScale = TRUE, group = 1)

Arguments

rmsea
The vector of RMSEA values to be plotted

n
Sample size of a dataset

df
Model degrees of freedom

ptile
The percentile rank of the distribution of the first RMSEA that users wish to plot a vertical line in the resulting graph

caption
The name vector of each element of rmsea

rmseaScale
If TRUE, the RMSEA scale is used in the x-axis. If FALSE, the chi-square scale is used in the x-axis.

group
The number of group that is used to calculate RMSEA.

Details

This function creates overlappling plots of the sampling distribution of RMSEA based on noncentral χ^2 distribution (MacCallum, Browne, & Suguwara, 1996). First, the noncentrality parameter (λ) is calculated from RMSEA (Steiger, 1998; Dudgeon, 2004) by

λ = (N - 1)d\varepsilon^2 / K,

where N is sample size, d is the model degree of freedom, K is the number of group, and \varepsilon is the population RMSEA. Next, the noncentral χ^2 distribution with a specified df and noncentrality parameter is plotted. Thus, the x-axis represents the sample χ^2 value. The sample χ^2 value can be transformed to the sample RMSEA scale (\hat{\varepsilon}) by

\hat{\varepsilon} = √{K}√{\frac{χ^2 - d}{(N - 1)d}},

where χ^2 is the χ^2 value obtained from the noncentral χ^2 distribution.

Author(s)

Sunthud Pornprasertmanit (psunthud@gmail.com)

References

Dudgeon, P. (2004). A note on extending Steiger's (1998) multiple sample RMSEA adjustment to other noncentrality parameter-based statistic. Structural Equation Modeling, 11(3), 305–319. doi: 10.1207/s15328007sem1103_1

MacCallum, R. C., Browne, M. W., & Sugawara, H. M. (1996). Power analysis and determination of sample size for covariance structure modeling. Psychological Methods, 1(2), 130–149. doi: 10.1037/1082-989X.1.2.130

Steiger, J. H. (1998). A note on multiple sample extensions of the RMSEA fit index. Structural Equation Modeling, 5(4), 411–419. doi: 10.1080/10705519809540115

See Also

plotRMSEApower() to plot the statistical power based on population RMSEA given the sample size

findRMSEApower() to find the statistical power based on population RMSEA given a sample size

findRMSEAsamplesize() to find the minium sample size for a given statistical power based on population RMSEA

Examples

plotRMSEAdist(c(.05, .08), n = 200, df = 20, ptile = .95, rmseaScale = TRUE)
plotRMSEAdist(c(.05, .01), n = 200, df = 20, ptile = .05, rmseaScale = FALSE)

[Package semTools version 0.5-8 Index]