residualCovariate {semTools}	R Documentation

Residual-center all target indicators by covariates

Description

This function will regress target variables on the covariate and replace the target variables by the residual of the regression analysis. This procedure is useful to control the covariate from the analysis model (Geldhof, Pornprasertmanit, Schoemann, & Little, 2013).

Usage

residualCovariate(data, targetVar, covVar)

Arguments

data
The desired data to be transformed.

targetVar
Varible names or the position of indicators that users wish to be residual centered (as dependent variables)

covVar
Covariate names or the position of the covariates using for residual centering (as independent variables) onto target variables

Value

The data that the target variables replaced by the residuals

Author(s)

Sunthud Pornprasertmanit (psunthud@gmail.com)

References

Geldhof, G. J., Pornprasertmanit, S., Schoemann, A. M., & Little, T. D. (2013). Orthogonalizing through residual centering: Extended applications and caveats. Educational and Psychological Measurement, 73(1), 27–46. doi: 10.1177/0013164412445473

See Also

indProd() For creating the indicator products with no centering, mean centering, double-mean centering, or residual centering.

Examples

dat <- residualCovariate(attitude, 2:7, 1)

[Package semTools version 0.5-8 Index]