semTools {semTools}	R Documentation

semTools: Useful Tools for Structural Equation Modeling

Description

The semTools package provides many miscellaneous functions that are useful for statistical analysis involving SEM in R. Many functions extend the funtionality of the lavaan package. Some sets of functions in semTools correspond to the same theme. We call such a collection of functions a suite. Our suites include:

Model Fit Evaluation: moreFitIndices(), nullRMSEA(), singleParamTest(), epcEquivFit(), and chisqSmallN()

Measurement Invariance: measEq.syntax(), partialInvariance(), partialInvarianceCat(), and permuteMeasEq()

Power Analysis: SSpower(), findRMSEApower(), plotRMSEApower(), plotRMSEAdist(), findRMSEAsamplesize(), findRMSEApowernested(), plotRMSEApowernested(), and findRMSEAsamplesizenested()

Missing Data Analysis: auxiliary(), twostage(), fmi(), bsBootMiss(), quark(), and combinequark()

Latent Interactions: indProd(), orthogonalize(), probe2WayMC(), probe3WayMC(), probe2WayRC(), probe3WayRC(), and plotProbe()

Exploratory Factor Analysis (EFA): efa.ekc()

Reliability Estimation: compRelSEM() and maximalRelia() (see also AVE())

Parceling: parcelAllocation(), PAVranking(), and poolMAlloc()

Non-Normality: skew(), kurtosis(), mardiaSkew(), mardiaKurtosis(), and mvrnonnorm()

All users of R (or SEM) are invited to submit functions or ideas for functions by contacting the maintainer, Terrence Jorgensen (TJorgensen314@gmail.com). Contributors are encouraged to use Roxygen comments to document their contributed code, which is consistent with the rest of semTools. Read the vignette from the roxygen2 package for details: vignette("rd", package = "roxygen2")

[Package semTools version 0.5-8 Index]