semTools-deprecated {semTools}	R Documentation

Deprecated functions in package semTools.

Description

The functions listed below are deprecated and will be defunct in the near future. When possible, alternative functions with similar functionality are also mentioned. Help pages for deprecated functions are available at help("semTools-deprecated").

Usage

reliability(object, what = c("alpha", "omega", "omega2", "omega3", "ave"),
  return.total = FALSE, dropSingle = TRUE, omit.factors = character(0),
  omit.indicators = character(0), omit.imps = c("no.conv", "no.se"))

reliabilityL2(object, secondFactor, omit.imps = c("no.conv", "no.se"))

Reliability

The original reliability function was suboptimally designed. For example, AVE was returned, which is not a reliability index. Also, alpha and several omega-type coefficients were returned, including the original formula that was inappropriate for models with complex structure. Some features could be controlled by the user for one but not both types of index For example, alpha for categorical indicators was returned on both the observed and latent-response scales, but this was not an option for any omega-type indices. The omegas differed in terms of whether the observed or model-implied covariance matrix was used in the denominator, but alpha was only computed using the observed matrix. These inconsistencies have been resolved in the new compRelSEM() function, which returns only one reliability index per composite, tailored to the user's requested features, for which there is much more flexibility. The average variance extracted is now available in a dedicated AVE() function.

Higher-Order Reliability

Originally, composite reliability of a single higher-order factor was estimated in a separate function: reliabilityL2. It is now available for multiple higher-order factors in the same model, and from the same compRelSEM() function that estimates reliability for first-order factors, using the higher= argument to name higher-order factors in the fitted model.

[Package semTools version 0.5-8 Index]