skew {semTools}	R Documentation

Finding skewness

Description

Finding skewness (g_{1}) of an object

Usage

skew(object, population = FALSE)

Arguments

object
A vector used to find a skewness

population
TRUE to compute the parameter formula. FALSE to compute the sample statistic formula.

Details

The skewness computed by default is g_{1}, the third standardized moment of the empirical distribution of object. The population parameter skewness γ_{1} formula is

γ_{1} = \frac{μ_{3}}{μ^{3/2}_{2}},

where μ_{i} denotes the i order central moment.

The skewness formula for sample statistic g_{1} is

g_{1} = \frac{k_{3}}{k^{2}_{2}},

where k_{i} are the i order k-statistic.

The standard error of the skewness is

Var(\hat{g}_1) = \frac{6}{N}

where N is the sample size.

Value

A value of a skewness with a test statistic if the population is specified as FALSE

Author(s)

Sunthud Pornprasertmanit (psunthud@gmail.com)

References

Weisstein, Eric W. (n.d.). Skewness. Retrieved from MathWorld–A Wolfram Web Resource: https://mathworld.wolfram.com/Skewness.html

See Also

kurtosis() Find the univariate excessive kurtosis of a variable

mardiaSkew() Find Mardia's multivariate skewness of a set of variables

mardiaKurtosis() Find the Mardia's multivariate kurtosis of a set of variables

Examples

skew(1:5)

[Package semTools version 0.5-8 Index]